+++ server/test/repositories/search.repository.mock.ts
@@ -3,8 +3,8 @@ import { ISearchRepository } from '@app/domain';
 export const newSearchRepositoryMock = (): jest.Mocked<ISearchRepository> => {
   return {
     init: jest.fn(),
-    searchAssets: jest.fn(),
-    searchCLIP: jest.fn(),
+    searchMetadata: jest.fn(),
+    searchSmart: jest.fn(),
     searchFaces: jest.fn(),
     upsert: jest.fn(),
   };
